backend/requirements.txt
flask>=2.0.3
python-dotenv>=0.19.2
authlib>=1.0
requests>=2.27.1
PyJWT>=2.8.0
cryptography>=3.4.8
flask-cors>=4.0.0
pymongo[srv]==4.3.3
google-generativeai>=0.3.0
python-jose>=3.3.0
flask-limiter>=3.0.0
flask-restx>=1.1.0